refactor: split report section composition
